backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.6.0
google-generativeai>=0.3.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
Pillow>=10.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
